partial fix

+++ src/models/action_encoder.py
@@ -104,13 +104,13 @@ class ActionEncoder(nn.Module):
                 kaiming_weights = torch.rand(weights.shape[0], input - weights.shape[1])
                 torch.nn.init.kaiming_uniform_(kaiming_weights, mode='fan_in', nonlinearity='leaky_relu')
 
-                weights = torch.cat([weights, kaiming_weights], dim=1)
+                weights = torch.cat([weights.double(), kaiming_weights.double()], dim=1)
 
             if output != weights.shape[0]:
                 kaiming_weights = torch.rand(output - weights.shape[0], input)
                 torch.nn.init.kaiming_uniform_(kaiming_weights, mode='fan_in', nonlinearity='leaky_relu')
 
-                weights = torch.cat([weights, kaiming_weights], dim=0)
+                weights = torch.cat([weights.double(), kaiming_weights.double()], dim=0)
 
             # Set
             layer.weight = nn.Parameter(weights)
